% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRegGraph.R
\name{makeRegulatorGraph}
\alias{makeRegulatorGraph}
\title{produce a graph that summarizes regulator-module relationships}
\usage{
makeRegulatorGraph(dataFolder = ".", modpatt = "^Module_")
}
\arguments{
\item{dataFolder}{character(1) specifying location of regulators 
and gene_expression CSV files by module, with filenames of form
'Module_[n]_...' by default; other filename formats can be
specified by modpatt parameter.  `_gene_expression.csv` and
`_regulators.csv` are the assumed suffixes.}

\item{modpatt}{character(1) regular expression used to pick out the CSV files}
}
\value{
instance of graph::graphNEL, mode 'directed'
}
\description{
produce a graph that summarizes regulator-module relationships
}
\note{
The returned graph has regulators, module names,
and module gene symbols as nodes.  A module has edges to
all member genes, and a regulator has edges to all modules
that it regulates.
}
\examples{
data(regulGBM)
regulGBM
}
